requirements.txt
groq
fastapi
uvicorn
python-dotenv
sounddevice
soundfile
pydub
pyaudio
streamlit